This is for some minor issues I found in alloc

+++ src/clibs/alloc/aalloc.c
@@ -1,47 +1,47 @@
 /* Software License Agreement
- * 
+ *
  *     Copyright(C) 1994-2018 David Lindauer, (LADSoft)
- * 
+ *
  *     This file is part of the Orange C Compiler package.
- * 
+ *
  *     The Orange C Compiler package is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
  *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version, with the addition of the 
+ *     (at your option) any later version, with the addition of the
  *     Orange C "Target Code" exception.
- * 
+ *
  *     The Orange C Compiler package is distributed in the hope that it will be useful,
  *     but WITHOUT ANY WARRANTY; without even the implied warranty of
  *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *     GNU General Public License for more details.
- * 
+ *
  *     You should have received a copy of the GNU General Public License
  *     along with Orange C.  If not, see <http://www.gnu.org/licenses/>.
- * 
+ *
  *     contact information:
  *         email: TouchStone222@runbox.com <David Lindauer>
- * 
+ *
  */
 
 #include <stdlib.h>
 #include <string.h>
 
-void *_RTL_FUNC aligned_alloc(size_t __align, size_t __size)
+void* _RTL_FUNC aligned_alloc(size_t __align, size_t __size)
 {
     // this is brute force and uses too much memory...
     // i don't have time to rewrite the entire allocation algorithm though...
-    int sz = __align + __size;
-    void *t = malloc(sz);
+    size_t sz = __align + __size;
+    void* t = malloc(sz);
     if (t)
     {
         int ofs = (unsigned)t % __align;
         if (ofs)
         {
             ofs = __align - ofs;
-            t = (char *)t + ofs;
-            *(unsigned *)((char *)t-sizeof(unsigned)) = - ofs; /// stuff in a negative value into what would normally
-                                                                // be the size field, to signify this
-                                                                // was aligned to free()
+            t = (char*)t + ofs;
+            *(unsigned*)((char*)t - sizeof(unsigned)) = -ofs;  /// stuff in a negative value into what would normally
+                                                               // be the size field, to signify this
+                                                               // was aligned to free()
         }
     }
     return t;
